requirements.txt
PyQt5
dropbox
openai
python-dotenv
PyPDF2
